requirements.txt
streamlit
python-dotenv
groq
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-huggingface
langchain-openai